refactor(web): extract the experiment data table into a reusable DataTable
