+++ apps/web/components/data-table/cells/user/data-table-user-cell.tsx
@@ -1,0 +1,21 @@
+"use client";
+
+import { ContributorIdentity } from "../../../contributor/contributor-identity";
+
+interface ExperimentDataTableUserCellProps {
+  data: string;
+  columnName: string;
+}
+
+/**
+ * Thin wrapper preserved for the existing call sites (experiment data
+ * table, dashboard table widget). Rendering lives in `ContributorIdentity`
+ * so the same look applies in filter dropdowns, filter chips, and any
+ * future CONTRIBUTOR surface.
+ */
+export function DataTableUserCell({
+  data,
+  columnName: _columnName,
+}: ExperimentDataTableUserCellProps) {
+  return <ContributorIdentity data={data} />;
+}
