+++ apps/web/components/experiment-data/experiment-data-table.test.tsx
@@ -25,15 +25,15 @@ vi.mock("./table-chart/experiment-data-table-chart", () => ({
   ExperimentDataTableChart: () => <div>Chart</div>,
 }));
 
-vi.mock("~/components/experiment-data/experiment-data-utils", () => ({
-  ExperimentTableHeader: ({ headerGroups }: { headerGroups: unknown[] }) => (
+vi.mock("~/components/data-table/data-table-utils", () => ({
+  DataTableHeader: ({ headerGroups }: { headerGroups: unknown[] }) => (
     <thead>
       <tr>
         <th>Header ({headerGroups.length} groups)</th>
       </tr>
     </thead>
   ),
-  ExperimentDataRows: ({ rows }: { rows: unknown[] }) => (
+  DataTableRows: ({ rows }: { rows: unknown[] }) => (
     <tr>
       <td>{rows.length} data rows</td>
     </tr>
@@ -109,7 +109,7 @@ describe("ExperimentDataTable", () => {
     setupHook();
     render(<ExperimentDataTable {...defaultProps} />);
     expect(screen.getByText("Test Table")).toBeInTheDocument();
-    expect(screen.getByText(/experimentDataTable.totalRows.*50/)).toBeInTheDocument();
+    expect(screen.getByText(/dataTable.totalRows.*50/)).toBeInTheDocument();
     expect(screen.getByText("2 data rows")).toBeInTheDocument();
   });
 
@@ -131,10 +131,10 @@ describe("ExperimentDataTable", () => {
     render(<ExperimentDataTable {...defaultProps} />);
 
     const user = userEvent.setup();
-    await user.click(screen.getByTitle("experimentDataTable.next"));
+    await user.click(screen.getByTitle("dataTable.next"));
     expect(mockUseExperimentData).toHaveBeenLastCalledWith(expect.objectContaining({ page: 2 }));
 
-    await user.click(screen.getByTitle("experimentDataTable.previous"));
+    await user.click(screen.getByTitle("dataTable.previous"));
     expect(mockUseExperimentData).toHaveBeenLastCalledWith(expect.objectContaining({ page: 1 }));
   });
 
@@ -142,11 +142,8 @@ describe("ExperimentDataTable", () => {
     setupHook({ tableMetadata: { ...mockTableMetadata, totalPages: 1, totalRows: 5 } });
     render(<ExperimentDataTable {...defaultProps} />);
 
-    expect(screen.getByTitle("experimentDataTable.previous")).toHaveAttribute(
-      "aria-disabled",
-      "true",
-    );
-    expect(screen.getByTitle("experimentDataTable.next")).toHaveAttribute("aria-disabled", "true");
+    expect(screen.getByTitle("dataTable.previous")).toHaveAttribute("aria-disabled", "true");
+    expect(screen.getByTitle("dataTable.next")).toHaveAttribute("aria-disabled", "true");
   });
 
   it("navigates pages via Arrow keys", async () => {
@@ -174,8 +171,6 @@ describe("ExperimentDataTable", () => {
   it("displays page info text", () => {
     setupHook();
     render(<ExperimentDataTable {...defaultProps} />);
-    expect(
-      screen.getByText(/experimentDataTable.page.*1.*experimentDataTable.pageOf.*5/),
-    ).toBeInTheDocument();
+    expect(screen.getByText(/dataTable.page.*1.*dataTable.pageOf.*5/)).toBeInTheDocument();
   });
 });
